requirements.txt
#All requirements are listed here
langchain-chroma>=0.0.15
langchain-community>=0.0.15
ollama>=0.1.0
langchain-ollama>=0.0.15
gradio>=3.39.1
pypdf>=2.6.0    
langchain>=0.0.215